# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Agent-API
  version: "0.1.0"
  description: >
    Generischer API-Layer für Agenten (Mensch und KI) auf einen Baum aus
    Markdown-/JSON-/PII-JSON-/Binärdateien. Eine Instanz verwaltet genau
    einen Baum (kein Multi-Tenant-Modell, siehe spec.md §12). Grundlage:
    docs/spec.md.
  license:
    name: proprietary
servers:
  - url: /api/v1
    description: Basis-Pfad. Kein Tenant-Segment — eine Instanz je Baum/Projekt (spec.md §12).
security:
  - clientApiKey: []
    userToken: []
  - clientApiKey: []
...
paths:
...
  "/file/{path}":
    parameters:
      - $ref: "#/components/parameters/FilePath"
    get:
      operationId: getFile
      tags: [Tree]
      summary: Datei lesen
      description: >
        Text-Kinds (md/json/pii.json): Volltext + Version im Response-Body.
        `binary`: nur Metadaten, keine Bytes — dafür GET .../content (spec.md §4).
      responses:
        "200":
          description: Metadaten (+ Inhalt bei Text-Kinds)
          headers:
            ETag: { schema: { type: string }, description: "= version, für If-None-Match-Caching" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileResponse" }
        "304": { description: "Not Modified (If-None-Match traf zu)" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
    post:
      operationId: writeFile
      tags: [Files]
      summary: Datei vollständig schreiben (Full Write / Upload)
      description: >
        Neue Datei anlegen oder vollständig ersetzen. Text-Kinds über
        `application/json`, `binary` über `multipart/form-data` (spec.md §5).
        `if_version` fehlt/leer bei Neuanlage, sonst Pflicht.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FullWriteTextRequest" }
          multipart/form-data:
            schema: { $ref: "#/components/schemas/FullWriteBinaryRequest" }
      responses:
        "200":
          description: Geschrieben
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriteResult" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/VersionConflict" }
    delete:
      operationId: deleteFile
      tags: [Files]
      summary: Datei löschen
      description: Immer explizit, immer mit `if_version` (spec.md §5).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeleteRequest" }
      responses:
        "204": { description: Gelöscht }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/VersionConflict" }
...
  "/file/{path}/edit":
    parameters:
      - $ref: "#/components/parameters/FilePath"
    post:
      operationId: editFile
      tags: [Files]
      summary: str_replace-Patch (nur Text-Kinds)
      description: >
        `old_str` muss genau einmal im aktuellen Inhalt vorkommen (0 oder
        >1 Treffer = 400). Doppelte Absicherung: Versions-Check UND exakter
        String-Match (spec.md §5).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EditRequest" }
      responses:
        "200":
          description: Geändert
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriteResult" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/VersionConflict" }
  "/file/{path}/append":
    parameters:
      - $ref: "#/components/parameters/FilePath"
    post:
      operationId: appendFile
      tags: [Files]
      summary: An Datei anhängen (nur Text-Kinds, kein old_str nötig)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppendRequest" }
      responses:
        "200":
          description: Angehängt
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriteResult" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/VersionConflict" }
...
  /system/clients:
    get:
      operationId: listClients
      tags: [Clients]
      summary: Registrierte Clients auflisten (Admin-only)
      responses:
        "200":
          description: Liste, ohne geheime Felder (api_key_hash, private Schlüssel — es gibt keine, nur public_key wird gespeichert)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClientRegistryEntryPublic" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createClient
      tags: [Clients]
      summary: Neuen Client registrieren (Admin-only)
      description: Der erzeugte `api_key` wird nur in dieser Antwort im Klartext zurückgegeben, danach nur noch als Hash gespeichert.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientRegistryEntryCreate" }
      responses:
        "201":
          description: Erstellt
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientRegistryEntryCreated" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  "/system/clients/{client_id}":
    parameters:
      - name: client_id
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getClient
      tags: [Clients]
      summary: Einzelnen Client-Registry-Eintrag lesen (Admin-only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientRegistryEntryPublic" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateClient
      tags: [Clients]
      summary: Client-Metadaten ändern (Admin-only)
      description: z.B. `issues_user_tokens` umschalten. Schlüsselrotation und Widerruf laufen über eigene Endpunkte.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientRegistryEntryUpdate" }
      responses:
        "200":
          description: Geändert
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientRegistryEntryPublic" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: revokeClient
      tags: [Clients]
      summary: Client widerrufen (Admin-only, soft — kein Hard-Delete)
      description: "Setzt `revoked: true`. Bereits ausgestellte User-Tokens bleiben bis `exp` gültig (max. 15 min, spec.md §10)."
      responses:
        "204": { description: Widerrufen }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  "/system/clients/{client_id}/rotate-api-key":
    parameters:
      - name: client_id
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: rotateClientApiKey
      tags: [Clients]
      summary: API-Key eines Clients rotieren (Admin-only)
      description: Alter Key wird sofort ungültig. Neuer Key nur in dieser Antwort im Klartext.
      responses:
        "200":
          description: Rotiert
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyIssued" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  "/system/clients/{client_id}/signing-keys":
    parameters:
      - name: client_id
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: addSigningKey
      tags: [Clients]
      summary: Neuen Public Key für User-Token-Verifikation hinzufügen (Admin-only)
      description: >
        Rotation ohne Downtime: der Client generiert sein eigenes
        Ed25519-Schlüsselpaar und registriert nur den Public Key hier — der
        private Schlüssel verlässt den Client nie (spec.md §10/§15 im
        Design-Dokument).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SigningKeyCreate" }
      responses:
        "201":
          description: Hinzugefügt
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SigningKey" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  "/system/clients/{client_id}/signing-keys/{kid}":
    parameters:
      - name: client_id
        in: path
        required: true
        schema: { type: string }
      - name: kid
        in: path
        required: true
        schema: { type: string }
    delete:
      operationId: revokeSigningKey
      tags: [Clients]
      summary: Public Key widerrufen (Admin-only, soft)
      description: Setzt `revoked_at`. Bereits ausgestellte Tokens mit diesem `kid` werden ab sofort abgelehnt.
      responses:
        "204": { description: Widerrufen }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    clientApiKey:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <client_api_key>` — beweist den client_id
        (spec.md §10a). Auf JEDEM Call erforderlich.
    userToken:
      type: apiKey
      in: header
      name: X-User-Token
      description: >
        Signierter Kurzzeit-JWT (EdDSA, ~15 min gültig) — beweist den
        user_id (spec.md §10b). Erforderlich bei jedem Call, AUSSER der
        Client hat `fixed_user_id` in der Registry gesetzt (autonomer
        Agent ohne Menschen dahinter) — dieser Fall ist in `security:`
        oben als zweite, alternative Anforderung (nur clientApiKey)
        hinterlegt. Welcher der beiden Fälle zutrifft, weiß der Server
        erst zur Laufzeit aus der Client-Registry — bei fälschlich
        fehlendem Token für einen nicht-autonomen Client lehnt er mit
        401 ab.
  parameters:
    FilePath:
      name: path
      in: path
      required: true
      schema: { type: string }
      description: >
        Relativer POSIX-Pfad zur Datei INKLUSIVE Schrägstrichen, z.B.
        `lv/sanitaer.md`. Wird server-seitig als Greedy-Match behandelt
        (FastAPI `path`-Konverter) — Clients dürfen den Pfad nicht
        segmentweise urlencodieren.
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema: { type: string }
      description: Für Caching — entspricht der zuletzt gesehenen `version`.
  schemas:
    Kind:
      type: string
      enum: [dir, md, json, "pii.json", binary]
...
    FileResponse:
      type: object
      required: [path, kind, version]
      properties:
        path: { type: string }
        kind: { $ref: "#/components/schemas/Kind" }
        version: { type: string }
        size: { type: integer }
        mime_type: { type: string, description: "nur kind=binary" }
        content: { type: string, description: "nur Text-Kinds, Volltext" }
    WriteResult:
      type: object
      required: [path, version, commit_id]
      properties:
        path: { type: string }
        version: { type: string, description: "neue Version nach dem Schreiben" }
        commit_id: { type: string, description: "Git-Commit-SHA dieser Transaktion" }
    EditRequest:
      type: object
      required: [old_str, new_str, if_version, reason]
      properties:
        old_str: { type: string }
        new_str: { type: string }
        if_version: { type: string }
        reason: { type: string, minLength: 1 }
        task_id: { type: string }
    AppendRequest:
      type: object
      required: [content, if_version, reason]
      properties:
        content: { type: string }
        if_version: { type: string }
        reason: { type: string, minLength: 1 }
        task_id: { type: string }
    FullWriteTextRequest:
      type: object
      required: [content, reason]
      properties:
        content: { type: string }
        if_version: { type: string, description: "weglassen bei Neuanlage" }
        reason: { type: string, minLength: 1 }
        task_id: { type: string }
    FullWriteBinaryRequest:
      type: object
      required: [file, reason]
      properties:
        file: { type: string, contentEncoding: binary, description: "Rohbytes, multipart" }
        if_version: { type: string, description: "= Hash der bisherigen Bytes; weglassen bei Neuanlage" }
        reason: { type: string, minLength: 1 }
        task_id: { type: string }
    DeleteRequest:
      type: object
      required: [if_version, reason]
      properties:
        if_version: { type: string }
        reason: { type: string, minLength: 1 }
        task_id: { type: string }
...
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: "Maschinenlesbarer Code, z.B. unauthorized, acl_denied, not_found, no_match, ambiguous_match" }
        message: { type: string }
        details: { type: object }
    ConflictError:
      type: object
      required: [error, current_version]
      properties:
        error: { type: string, const: version_conflict }
        message: { type: string }
        current_content: { type: string, nullable: true, description: "null bei kind=binary" }
        current_version: { type: string }

    # --- Client-Registry (spec.md §10, Design-Doc §14/§15) ---
    ClientType:
      type: string
      enum: [bff, orchestrator, autonomous-agent]
    SigningKey:
      type: object
      required: [kid, public_key, algorithm, active]
      properties:
        kid: { type: string }
        public_key: { type: string }
        algorithm: { type: string, const: EdDSA }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time, nullable: true }
    SigningKeyCreate:
      type: object
      required: [kid, public_key]
      properties:
        kid: { type: string }
        public_key: { type: string }
        algorithm: { type: string, const: EdDSA, default: EdDSA }
    ClientRegistryEntryPublic:
      type: object
      required: [client_id, type, issues_user_tokens, revoked, created_at]
      description: Öffentliche Sicht — enthält NIE api_key_hash oder private Schlüssel (es werden ohnehin nur Public Keys gespeichert).
      properties:
        client_id: { type: string }
        type: { $ref: "#/components/schemas/ClientType" }
        issues_user_tokens: { type: boolean }
        signing_keys:
          type: array
          items: { $ref: "#/components/schemas/SigningKey" }
        fixed_user_id: { type: string, nullable: true, description: "nur bei type=autonomous-agent gesetzt" }
        revoked: { type: boolean }
        created_at: { type: string, format: date-time }
    ClientRegistryEntryCreate:
      type: object
      required: [client_id, type]
      properties:
        client_id: { type: string }
        type: { $ref: "#/components/schemas/ClientType" }
        issues_user_tokens: { type: boolean, default: false }
        fixed_user_id: { type: string, description: "Pflicht bei type=autonomous-agent, sonst unzulässig" }
    ClientRegistryEntryUpdate:
      type: object
      properties:
        issues_user_tokens: { type: boolean }
    ApiKeyIssued:
      type: object
      required: [client_id, api_key]
      properties:
        client_id: { type: string }
        api_key: { type: string, description: "Klartext — wird nur hier einmalig zurückgegeben" }
    ClientRegistryEntryCreated:
      allOf:
        - { $ref: "#/components/schemas/ClientRegistryEntryPublic" }
        - type: object
          required: [api_key]
          properties:
            api_key: { type: string, description: "Klartext — wird nur hier einmalig zurückgegeben" }

    # --- Audit-Log (spec.md §9) ---
...
  responses:
    Unauthorized:
      description: "Client-API-Key fehlt/ungültig, oder User-Token fehlt/ungültig/abgelaufen wo nötig"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: "ACL verweigert (spec.md §8) oder Admin-Endpunkt ohne Admin-user_id"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Pfad/Ressource nicht gefunden, oder ACL blendet aus (Sichtbarkeit wie „nicht gefunden")
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: "Ungültige Anfrage (z.B. old_str 0 oder >1 Treffer, Pflichtfeld fehlt)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    VersionConflict:
      description: "if_version stimmt nicht mehr mit dem aktuellen Stand überein (spec.md §11)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ConflictError" }